+++ source/src/vm/pv2000/pv2000.cpp
@@ -300,7 +300,7 @@ bool VM::process_state(FILEIO* state_fio, bool loading)
  		// const char *name = typeid(*device).name();
 		//       But, using get_device_name() instead of typeid(foo).name() 20181008 K.O
 		const char *name = device->get_device_name();
-		int len = strlen(name);
+		int len = (int)strlen(name);
 		
 		if(!state_fio->StateCheckInt32(len)) {
 			if(loading) {
